Move landing page to screens, features to module
