+++ src/screens/LandingPage/LandingPage.jsx
@@ -1,36 +1,16 @@
 import { Button, Label, TextInput, Tooltip } from "flowbite-react";
-import InputField from "./Forms/InputField";
+import InputField from "../../components/Forms/InputField";
 import { InformationCircleIcon } from "@heroicons/react/24/solid";
 
 import { Field, Formik } from "formik";
 
 import { router } from "next/router";
-import { createOrganization } from "../queries/organizations";
+import { createOrganization } from "../../queries/organizations";
 import React, { useState } from "react";
-import { createOrganizationValidator } from "../screens/Onboarding/helpers";
-import Footer from "../screens/Events/Admin/Attendance/Footer";
-import AppFooter from "./Footer";
-
-const features = [
-  {
-    title: "Track your impact",
-    description:
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-    imageUrl: "/images/landing-feature-steps.svg",
-  },
-  {
-    title: "Plan memorable events",
-    description:
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-    imageUrl: "/images/landing-feature-schedule.svg",
-  },
-  {
-    title: "Track your impact",
-    description:
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-    imageUrl: "/images/landing-feature-steps.svg",
-  },
-];
+import { createOrganizationValidator } from "../Onboarding/helpers";
+import Footer from "../Events/Admin/Attendance/Footer";
+import AppFooter from "../../components/Footer";
+import { features } from "./features";
 
 const goToLogin = () => {
   router.push("/login");
