+++ src/screens/LandingPage/index.jsx
@@ -1,0 +1,3 @@
+import LandingPage from "./LandingPage";
+
+export default LandingPage;
